Raise a background sync event to sync reviews with server.
When a user submits a review, store the review data in the temp reviews-sync store and then raise a background sync event. When the user is then online, read the reviews to sync from the cache, post each to the server, remove the reviews from the temp sync store and add them to the normal reviews store.

Ensure the reviews-sync store Id is NOT used when POSTing to the server. We want the server to allocate the Id for the review.